# evidence/openapi.json
{"openapi":"3.1.0","info":{"title":"AI Workspace — SWE Review","version":"0.1.0"},"paths":{"/api/config":{"get":{"summary":"Config","operationId":"config_api_config_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/catalog":{"get":{"summary":"Catalog","operationId":"catalog_api_catalog_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/reviews":{"get":{"summary":"Reviews","operationId":"reviews_api_reviews_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}},"post":{"summary":"Create","operationId":"create_api_reviews_post","parameters":[{"name":"Idempotency-Key","in":"header","required":true,"schema":{"type":"string","minLength":1,"maxLength":100,"pattern":"^[a-zA-Z0-9_.:-]+$","title":"Idempotency-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewInput"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/reviews/{rid}":{"get":{"summary":"Read","operationId":"read_api_reviews__rid__get","parameters":[{"name":"rid","in":"path","required":true,"schema":{"type":"string","title":"Rid"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/reviews/{rid}/answers":{"post":{"summary":"Answer","operationId":"answer_api_reviews__rid__answers_post","parameters":[{"name":"rid","in":"path","required":true,"schema":{"type":"string","title":"Rid"}},{"name":"Idempotency-Key","in":"header","required":true,"schema":{"type":"string","minLength":1,"maxLength":100,"pattern":"^[a-zA-Z0-9_.:-]+$","title":"Idempotency-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnswersInput"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/reviews/{rid}/context":{"get":{"summary":"Context","operationId":"context_api_reviews__rid__context_get","parameters":[{"name":"rid","in":"path","required":true,"schema":{"type":"string","title":"Rid"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/reviews/{rid}/model-output":{"post":{"summary":"Submit","operationId":"submit_api_reviews__rid__model_output_post","parameters":[{"name":"rid","in":"path","required":true,"schema":{"type":"string","title":"Rid"}},{"name":"Idempotency-Key","in":"header","required":true,"schema":{"type":"string","minLength":1,"maxLength":100,"pattern":"^[a-zA-Z0-9_.:-]+$","title":"Idempotency-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Submission"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/reviews/{rid}/export":{"get":{"summary":"Export","operationId":"export_api_reviews__rid__export_get","parameters":[{"name":"rid","in":"path","required":true,"schema":{"type":"string","title":"Rid"}},{"name":"format","in":"query","required":false,"schema":{"type":"string","pattern":"^(json|markdown)$","default":"markdown","title":"Format"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AnswersInput":{"properties":{"revision":{"type":"integer","minimum":0.0,"title":"Revision"},"answers":{"additionalProperties":{"type":"string"},"type":"object","maxProperties":3,"minProperties":1,"title":"Answers"}},"additionalProperties":false,"type":"object","required":["revision","answers"],"title":"AnswersInput"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ReviewInput":{"properties":{"mode":{"type":"string","enum":["mcp","scripted"],"title":"Mode","default":"mcp"},"workbench_record_id":{"type":"string","maxLength":100,"minLength":1,"title":"Workbench Record Id"},"title":{"type":"string","maxLength":120,"minLength":1,"title":"Title"},"design":{"type":"string","maxLength":8000,"minLength":10,"title":"Design"},"check_ids":{"items":{"type":"string"},"type":"array","maxItems":8,"minItems":1,"title":"Check Ids"}},"additionalProperties":false,"type":"object","required":["workbench_record_id","title","design","check_ids"],"title":"ReviewInput"},"Submission":{"properties":{"revision":{"type":"integer","minimum":0.0,"title":"Revision"},"input_sha256":{"type":"string","pattern":"^[a-f0-9]{64}$","title":"Input Sha256"},"output":{"additionalProperties":true,"type":"object","title":"Output"}},"additionalProperties":false,"type":"object","required":["revision","input_sha256","output"],"title":"Submission"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}